fix spacing and swiper issue
